Kindles are rude
* Today I learned part of my problem with kindle networking is that they
  don't set any flags when transitioning states, so there is no marker
to key a restart on. Went with the mindset of "if you're turning off
airoplane, you want networking on if you're on a kindle" and am using a
custom call to make sure networking enables correctly. Might apply that
to everyone, but for now kobo still works and kindle is consistant at
last.